move get_points_from_regions into plotter
